Create SongList
The assignment says:
Design a class SongList based on the UML diagram.
SongList contains a song array which can hold a MAX
(a static final field to specify the length of the songList
array to be created) number of songs. The size field is
used to keep track of how many songs have been added
into the songList so far. For the sorting method you can
either modify the sorting algorithm we have discussed
or you can implement your own.

UML:
SongList
————
- MAX: int
- songList: Song [ ]
- size: int
————
+SongList():
+addSong(Song): boolean
+findSong(String title): int
+swapSong(int, int):
+sortByWriter():
+sortByDuration():
+sortByGenre():
+sortByYear();
+toString(): String
+getSize(): int